image_downloader/requirements.txt
Django==4.2.6
python-decouple==3.6
psycopg2-binary==2.9.9
djangorestframework==3.15.1
